+++ core/src/main/java/tech/ydb/core/grpc/GrpcRequestSettings.java
@@ -1,6 +1,10 @@
 package tech.ydb.core.grpc;
 
+import java.util.function.Consumer;
+
 import javax.annotation.ParametersAreNonnullByDefault;
+
+import io.grpc.Metadata;
 
 /**
  * @author Nikolay Perfilov
@@ -8,10 +12,14 @@ import javax.annotation.ParametersAreNonnullByDefault;
 public class GrpcRequestSettings {
     private final long deadlineAfter;
     private final EndpointInfo preferredEndpoint;
+    private final Metadata extraHeaders;
+    private final Consumer<Metadata> trailersHandler;
 
     private GrpcRequestSettings(Builder builder) {
         this.deadlineAfter = builder.getDeadlineAfter();
         this.preferredEndpoint = builder.getPreferredEndpoint();
+        this.extraHeaders = builder.getExtraHeaders();
+        this.trailersHandler = builder.getTrailersHandler();
     }
 
     public static Builder newBuilder() {
@@ -26,10 +34,20 @@ public class GrpcRequestSettings {
         return preferredEndpoint;
     }
 
+    public Metadata getExtraHeaders() {
+        return extraHeaders;
+    }
+
+    public Consumer<Metadata> getTrailersHandler() {
+        return trailersHandler;
+    }
+
     @ParametersAreNonnullByDefault
     public static final class Builder {
         private long deadlineAfter = 0;
         private EndpointInfo preferredEndpoint = null;
+        private Metadata extraHeaders = null;
+        private Consumer<Metadata> trailersHandler = null;
 
         public Builder withDeadlineAfter(long deadlineAfter) {
             this.deadlineAfter = deadlineAfter;
@@ -41,12 +59,30 @@ public class GrpcRequestSettings {
             return this;
         }
 
+        public Builder withExtraHeaders(Metadata headers) {
+            this.extraHeaders = headers;
+            return this;
+        }
+
+        public Builder withTrailersHandler(Consumer<Metadata> handler) {
+            this.trailersHandler = handler;
+            return this;
+        }
+
         public long getDeadlineAfter() {
             return deadlineAfter;
         }
 
         public EndpointInfo getPreferredEndpoint() {
             return preferredEndpoint;
+        }
+
+        public Metadata getExtraHeaders() {
+            return extraHeaders;
+        }
+
+        public Consumer<Metadata> getTrailersHandler() {
+            return trailersHandler;
         }
 
         public GrpcRequestSettings build() {