+++ core/src/main/java/tech/ydb/core/grpc/impl/grpc/GrpcTransportImpl.java
@@ -86,7 +86,7 @@ public class GrpcTransportImpl extends GrpcTransport {
 
     /**
      * Establish connection for grpc channel(s) if its currently IDLE
-     * Returns a future to a first {@link ConnectivityState} that is not IDLE or CONNECTING
+     * @return a future to a first {@link ConnectivityState} that is not IDLE or CONNECTING
      */
     public CompletableFuture<ConnectivityState> tryToConnect() {
         CompletableFuture<ConnectivityState> promise = new CompletableFuture<>();
@@ -201,7 +201,7 @@ public class GrpcTransportImpl extends GrpcTransport {
         if (logger.isDebugEnabled()) {
             logger.debug("Sending request to {}, method `{}', request: `{}'", channel.authority(), method, request );
         }
-        sendOneRequest(call, request, new UnaryStreamToFuture<>(promise));
+        sendOneRequest(call, request, settings, new UnaryStreamToFuture<>(promise, settings.getTrailersHandler()));
         return promise;
     }
 
@@ -216,7 +216,8 @@ public class GrpcTransportImpl extends GrpcTransport {
         if (logger.isDebugEnabled()) {
             logger.debug("Sending request to {}, method `{}', request: `{}'", channel.authority(), method, request );
         }
-        sendOneRequest(call, request, new UnaryStreamToConsumer<>(consumer));
+        sendOneRequest(call, request, settings,
+                new UnaryStreamToConsumer<>(consumer, null, settings.getTrailersHandler()));
     }
 
     @Override
@@ -230,7 +231,8 @@ public class GrpcTransportImpl extends GrpcTransport {
         if (logger.isDebugEnabled()) {
             logger.debug("Sending request to {}, method `{}', request: `{}'", channel.authority(), method, request );
         }
-        sendOneRequest(call, request, new UnaryStreamToBiConsumer<>(consumer));
+        sendOneRequest(call, request, settings,
+                new UnaryStreamToBiConsumer<>(consumer, null, settings.getTrailersHandler()));
     }
 
     @Override
@@ -241,7 +243,7 @@ public class GrpcTransportImpl extends GrpcTransport {
             GrpcRequestSettings settings,
             StreamObserver<RespT> observer) {
         ClientCall<ReqT, RespT> call = channel.newCall(method, callOptions);
-        sendOneRequest(call, request, new ServerStreamToObserver<>(observer, call));
+        sendOneRequest(call, request, settings, new ServerStreamToObserver<>(observer, call));
         return () -> {
             call.cancel("Cancelled on user request", new CancellationException());
         };
@@ -257,8 +259,9 @@ public class GrpcTransportImpl extends GrpcTransport {
         AsyncBidiStreamingOutAdapter<ReqT, RespT> adapter
                 = new AsyncBidiStreamingOutAdapter<>(call);
         AsyncBidiStreamingInAdapter<ReqT, RespT> responseListener
-                = new AsyncBidiStreamingInAdapter<>(observer, adapter);
-        call.start(responseListener, new Metadata());
+                = new AsyncBidiStreamingInAdapter<>(observer, adapter, null, settings.getTrailersHandler());
+        Metadata extra = settings.getExtraHeaders();
+        call.start(responseListener, extra != null ? extra :new Metadata());
         responseListener.onStart();
         return adapter;
     }